refactoring and cleaning

+++ dbinit/db_init.py
@@ -3,22 +3,15 @@ from langchain_community.document_loaders.pdf import PyMuPDFLoader
 from langchain_community.embeddings.sentence_transformer import SentenceTransformerEmbeddings
 from langchain_community.vectorstores import Chroma
 from langchain_text_splitters import CharacterTextSplitter
-
 import uuid
 import chromadb
 from chromadb.config import DEFAULT_TENANT, DEFAULT_DATABASE, Settings
 
-# Load documents
 loader = DirectoryLoader("documents", glob="*.pdf", loader_cls=PyMuPDFLoader)
 documents = loader.load()
 
-# Split documents into chunks
 text_splitter = CharacterTextSplitter(chunk_size=256, chunk_overlap=40)
 chunks = text_splitter.split_documents(documents)
-
-embedding_function = SentenceTransformerEmbeddings(model_name="all-MiniLM-L6-v2")
-
-# Creation of Chroma vectors db
 
 chroma_client = chromadb.HttpClient(
     host="rag-backend_chroma",
@@ -28,11 +21,13 @@ chroma_client = chromadb.HttpClient(
     database=DEFAULT_DATABASE
 )
 
+existing_collections = [collection.name for collection in chroma_client.list_collections()]
 try:
-    chroma_client.get_collection("data_collection")
-    chroma_client.delete_collection("data_collection")
-except:
-    pass
+    if "data_collection" in existing_collections:
+        chroma_client.delete_collection("data_collection")
+except Exception as e:
+    print(f"Warning: Could not delete collection. {e}")
+
 collection = chroma_client.create_collection("data_collection")
 
 for chunk in chunks:
@@ -41,6 +36,8 @@ for chunk in chunks:
         metadatas=chunk.metadata, 
         documents=chunk.page_content
     )
+
+embedding_function = SentenceTransformerEmbeddings(model_name="all-MiniLM-L6-v2")
 
 db = Chroma(
     client=chroma_client,
